requirements.txt
openai>=1.40.0
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.23.0
tqdm>=4.66.0
